Add support for importing Azure resources using azure-native
