--- pulumi-cloud-import/Pulumi.yaml
@@ -1,6 +1,0 @@
-name: pulumi-cloud-import
-runtime:
-  name: go
-  options:
-    binary: importer
-description: A minimal Pulumi YAML program
